some simplifications and fixes
